feat(hair): self-shadow, marschner model and flow map support (#1225)
A simple hair self-shadow based on a 4-step screen space raymarch, so the hair won't overglow with transmission.
Marschner model is a more physically based hair shading model compared to the default kajiya-kay model used in hair specular. User can now freely choose one of them. This model is dependent on self-shadow as it simulates both reflection and transmission.
Added support for hair flow map (or directional map). The texture needs to be put in the slot 7 (starting from 0) in the texture set, and back lighting flag has to be enabled in hair nif. Hair flow map is important for both of the hair models as they are anisotropic. The texture uses the RG channels to record the direction of hair strands in UV space. The direction always points from the hair tip to the root.